Implements data injectors
These are akin to materializers. They get passed into the materialize()
function. They are called with `.from_` and take in slightly different
arguments -- rather than dependencies and builder (which are not
necessary for materializers), they take in a target (the node to load
them into), and the kwargs.

These delegate directly to the LoadFromDecorator -- a class that can
generate loader nodes. Currently there's a strange integration point --
the load_from decorators introduces strange, namespaced names to avoids
conflicts, this does away with that. In the future, this will create one
node (and log the output), but for now, this creates two nodes, just as
the load_from decorator does.  A node to load that + metadata, and a
node to return the result and filter out the metadata.

There is some duplicated code between these and to.xyz -- particularly
around dynamically generating classnames. This is complex so that we can
have auto-completion in jupyter notebooks (which is where these will
commonly be called from). At some point we need to ensure that this
shares more code with the decorators (which have similar dynamism in
attributes), but for now these are well-constructed and fine the way
they are.